feat: Update Product and Order models to support price per kg
- Changed the Quantity property in OrderProduct from int to decimal for better precision.
- Added IsPricePerKg boolean property to the Product entity.
- Updated SeedData to reflect changes in Product constructor.
- Refactored Order tests to accommodate new Product constructor and validation logic.
- Introduced ProductBuilder for easier test data generation.
- Updated GraphQL mutations to handle new Product properties.
- Created migrations to add IsPricePerKg column and rename it appropriately.

+++ src/ComandaX.Application/DTOs/OrderProductDto.cs
@@ -3,7 +3,7 @@ namespace ComandaX.Application.DTOs;
 public sealed record OrderProductDto
 {
     public OrderProductDto() { }
-    public OrderProductDto(Guid productId, Guid orderId, int quantity, decimal totalPrice)
+    public OrderProductDto(Guid productId, Guid orderId, decimal quantity, decimal totalPrice)
     {
         ProductId = productId;
         Quantity = quantity;
@@ -12,6 +12,6 @@ public sealed record OrderProductDto
     }
     public Guid OrderId { get; set; }
     public Guid ProductId { get; set; }
-    public int Quantity { get; set; }
+    public decimal Quantity { get; set; }
     public decimal TotalPrice { get; set; }
 }
