+++ src/ComandaX.Application/Handlers/Orders/Commands/CreateOrder/CreateOrderProductDto.cs
@@ -2,5 +2,5 @@ namespace ComandaX.Application.Handlers.Orders.Commands.CreateOrder;
 
 public sealed record CreateOrderProductDto(
     Guid ProductId,
-    int Quantity
+    decimal Quantity
 );
